Moved dependency solver tests to a separate file
Also, started to use test-framework and HUnit to write tests and
split them into groups (now there's just one group though).
