+++ elm-get.cabal
@@ -92,6 +92,7 @@ Test-suite unit-test
   Type:                exitcode-stdio-1.0
   Hs-Source-Dirs:      tests, src
   Main-is:             Tests.hs
+  Other-modules:       SolverTests
   Build-depends:       aeson,
                        aeson-pretty,
                        ansi-wl-pprint,
@@ -103,6 +104,7 @@ Test-suite unit-test
                        Elm >= 0.12.2,
                        filepath,
                        HTTP,
+                       HUnit,
                        http-client >= 0.3,
                        http-client-tls,
                        http-types,
@@ -113,5 +115,7 @@ Test-suite unit-test
                        pretty,
                        process,
                        resourcet,
+                       test-framework,
+                       test-framework-hunit,
                        text,
                        vector
